Center Home Page and remove buttons
It automatically checks validity in stages. First it checks the format,
then it sends it to the server. 500ms after the user stops typing or
pasting, it will run these checks. This is the same as  it used to be,
but takes out the need to press the buttons.

Also adds a loading spinner so the user knows things are happening.